requirements.txt
Django==1.7
Pillow==2.6.1
argparse==1.2.1
django-admin-bootstrapped==2.1.0
django-bootstrap3==4.11.0
django-smart-selects==1.0.9.3
tablib==0.10.0
wsgiref==0.1.2

*To install django-smart-selects do,

pip install git+https://github.com/digi604/django-smart-selects.git
